+++ test/core_test_app/app/components/endpoints.rb
@@ -4,6 +4,7 @@ class Endpoints < Netzke::Base
   action :multiple_arguments
   action :array_as_argument
   action :return_value
+  action :non_existing
 
   # this action is using generic endpoint callback with scope
   action :callback_and_scope
@@ -15,7 +16,7 @@ class Endpoints < Netzke::Base
 
   def configure(c)
     super
-    c.bbar = [:with_response, :no_response, :multiple_arguments, :array_as_argument, :callback_and_scope, :return_value]
+    c.bbar = [:with_response, :no_response, :multiple_arguments, :array_as_argument, :callback_and_scope, :return_value, :non_existing]
 
     # Alternative way of defining bbar:
     # c.docked_items = [{
@@ -46,5 +47,17 @@ class Endpoints < Netzke::Base
   endpoint :get_answer do |params,this|
     raise "params expected to be null" if !params.nil?
     this.netzke_set_result(42) # special method that passes a value as argument to callback function
+  end
+
+  endpoint :server_non_existing do |params, this|
+    # won't get here
+  end
+
+  def invoke_endpoint(ep, params, configs)
+    if ep == "server_non_existing"
+      ep = "non_existing_child__endpoint"
+    end
+
+    super ep, params, configs
   end
 end